refactor: change package name (race -> action)
